# apps/web/src/lib/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Evidentia API",
    "description": "Evidence platform for field media: signed Cloudinary ingestion, AI enrichment, verification and traceable reporting.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/search": {
      "post": {
        "tags": [
          "search"
        ],
        "summary": "Search",
        "description": "Hybrid evidence search. Natural language is parsed into hard filters (site, dates, media type,\nverification) plus text for lexical, semantic and visual retrieval, fused with RRF.",
        "operationId": "search_v1_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/search/similar": {
      "post": {
        "tags": [
          "search"
        ],
        "summary": "Similar",
        "operationId": "similar_v1_search_similar_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimilarRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SimilarResult"
                  },
                  "type": "array",
                  "title": "Response Similar V1 Search Similar Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/search/{query_id}/feedback": {
      "post": {
        "tags": [
          "search"
        ],
        "summary": "Feedback",
        "description": "Clicks / 'add to evidence' / dismissals, logged for ranking evaluation and tuning.",
        "operationId": "feedback_v1_search__query_id__feedback_post",
        "parameters": [
          {
            "name": "query_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Query Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssetState": {
        "type": "string",
        "enum": [
          "awaiting_upload",
          "uploaded",
          "registered",
          "metadata_ready",
          "dedup_checked",
          "analyzing",
          "indexed",
          "review_required",
          "ready",
          "expired",
          "failed_retryable",
          "failed_permanent"
        ],
        "title": "AssetState"
      },
      "AssetSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "site_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Site Id"
          },
          "resource_type": {
            "$ref": "#/components/schemas/ResourceType"
          },
          "state": {
            "$ref": "#/components/schemas/AssetState"
          },
          "state_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State Reason"
          },
          "original_filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Filename"
          },
          "format": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Format"
          },
          "bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bytes"
          },
          "width": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          },
          "height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "capture_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capture Time"
          },
          "capture_time_source": {
            "$ref": "#/components/schemas/ProvenanceSource"
          },
          "location_source": {
            "$ref": "#/components/schemas/ProvenanceSource"
          },
          "declared_activity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Declared Activity"
          },
          "caption": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Caption"
          },
          "top_activity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top Activity"
          },
          "top_activity_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top Activity Confidence"
          },
          "review_priority": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Priority"
          },
          "flags": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Flags"
          },
          "exact_duplicate_of_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exact Duplicate Of Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "thumb_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumb Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "site_id",
          "resource_type",
          "state",
          "state_reason",
          "original_filename",
          "format",
          "bytes",
          "width",
          "height",
          "duration_seconds",
          "capture_time",
          "capture_time_source",
          "location_source",
          "declared_activity",
          "caption",
          "top_activity",
          "top_activity_confidence",
          "review_priority",
          "flags",
          "exact_duplicate_of_id",
          "created_at"
        ],
        "title": "AssetSummary"
      },
...
      "FeedbackRequest": {
        "properties": {
          "asset_id": {
            "type": "string",
            "format": "uuid",
            "title": "Asset Id"
          },
          "action": {
            "$ref": "#/components/schemas/SearchFeedbackAction"
          },
          "rank": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rank"
          }
        },
        "type": "object",
        "required": [
          "asset_id",
          "action"
        ],
        "title": "FeedbackRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MatchedObservation": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "ontology_type": {
            "type": "string",
            "title": "Ontology Type"
          },
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "predicate": {
            "type": "string",
            "title": "Predicate"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "ontology_type",
          "subject",
          "predicate",
          "confidence",
          "status"
        ],
        "title": "MatchedObservation"
      },
...
      "ProvenanceSource": {
        "type": "string",
        "enum": [
          "exif",
          "device",
          "user",
          "upload",
          "site",
          "none"
        ],
        "title": "ProvenanceSource"
      },
...
      "ResourceType": {
        "type": "string",
        "enum": [
          "image",
          "video"
        ],
        "title": "ResourceType"
      },
...
      "SearchFeedbackAction": {
        "type": "string",
        "enum": [
          "click",
          "add_to_evidence",
          "dismiss"
        ],
        "title": "SearchFeedbackAction"
      },
      "SearchFilters": {
        "properties": {
          "site_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Site Ids"
          },
          "activities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Activities"
          },
          "date_from": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date From"
          },
          "date_to": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date To"
          },
          "media_type": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(image|video)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Media Type"
          },
          "verified_only": {
            "type": "boolean",
            "title": "Verified Only",
            "default": false
          },
          "needs_review": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Needs Review"
          },
          "people_present": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "People Present"
          }
        },
        "type": "object",
        "title": "SearchFilters"
      },
      "SearchPlan": {
        "properties": {
          "filters": {
            "$ref": "#/components/schemas/SearchFilters"
          },
          "semantic_text": {
            "type": "string",
            "title": "Semantic Text",
            "default": ""
          },
          "visual_text": {
            "type": "string",
            "title": "Visual Text",
            "default": ""
          },
          "planner": {
            "type": "string",
            "title": "Planner",
            "default": "rules"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "title": "SearchPlan"
      },
      "SearchRequest": {
        "properties": {
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "query": {
            "type": "string",
            "maxLength": 500,
            "title": "Query",
            "default": ""
          },
          "filters": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SearchFilters"
              },
              {
                "type": "null"
              }
            ]
          },
          "limit": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Limit",
            "default": 24
          },
          "use_llm_planner": {
            "type": "boolean",
            "title": "Use Llm Planner",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "project_id"
        ],
        "title": "SearchRequest"
      },
      "SearchResponse": {
        "properties": {
          "query_id": {
            "type": "string",
            "format": "uuid",
            "title": "Query Id"
          },
          "plan": {
            "$ref": "#/components/schemas/SearchPlan"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/SearchResult"
            },
            "type": "array",
            "title": "Results"
          },
          "retrievers": {
            "additionalProperties": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "object",
            "title": "Retrievers"
          },
          "took_ms": {
            "type": "integer",
            "title": "Took Ms"
          },
          "query_terms": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Query Terms",
            "description": "Content words of the query"
          },
          "complete_matches": {
            "type": "integer",
            "title": "Complete Matches",
            "description": "Results that contain every query word",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "query_id",
          "plan",
          "results",
          "retrievers",
          "took_ms"
        ],
        "title": "SearchResponse"
      },
      "SearchResult": {
        "properties": {
          "asset": {
            "$ref": "#/components/schemas/AssetSummary"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reasons"
          },
          "contributions": {
            "additionalProperties": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "object",
            "title": "Contributions"
          },
          "matched_observations": {
            "items": {
              "$ref": "#/components/schemas/MatchedObservation"
            },
            "type": "array",
            "title": "Matched Observations"
          },
          "video_span": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VideoSpan"
              },
              {
                "type": "null"
              }
            ]
          },
          "missing_terms": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Terms",
            "description": "Query words this asset's description does not contain"
          }
        },
        "type": "object",
        "required": [
          "asset",
          "score",
          "reasons",
          "contributions",
          "matched_observations"
        ],
        "title": "SearchResult"
      },
...
      "SimilarRequest": {
        "properties": {
          "asset_id": {
            "type": "string",
            "format": "uuid",
            "title": "Asset Id"
          },
          "limit": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Limit",
            "default": 12
          },
          "same_project": {
            "type": "boolean",
            "title": "Same Project",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "asset_id"
        ],
        "title": "SimilarRequest"
      },
      "SimilarResult": {
        "properties": {
          "asset": {
            "$ref": "#/components/schemas/AssetSummary"
          },
          "similarity": {
            "type": "number",
            "title": "Similarity"
          }
        },
        "type": "object",
        "required": [
          "asset",
          "similarity"
        ],
        "title": "SimilarResult"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "VideoSpan": {
        "properties": {
          "start_ms": {
            "type": "integer",
            "title": "Start Ms"
          },
          "end_ms": {
            "type": "integer",
            "title": "End Ms"
          },
          "label": {
            "type": "string",
            "title": "Label"
          }
        },
        "type": "object",
        "required": [
          "start_ms",
          "end_ms",
          "label"
        ],
        "title": "VideoSpan"
      },
...
    }
  }
}